requirements/base.txt
django>=5.2,<5.3
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
django-environ>=0.12
django-cors-headers>=4.7
pymongo>=4.15
structlog>=24.0.0
